backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.115.12
groq==0.22.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.54
langchain-groq==0.3.2
langsmith==0.3.32
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.2
tenacity==9.1.2
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
zstandard==0.23.0
